+++ packages/unlock-js/src/PublicLock/v6/index.js
@@ -10,6 +10,8 @@ import isLockManager from './isLockManager'
 import cancelAndRefund from './cancelAndRefund'
 import shareKey from './shareKey'
 import getLock from './getLock'
+import addLockManager from './addLockManager'
+import renounceLockManager from './renounceLockManager'
 
 import v4 from '../v4'
 
@@ -43,4 +45,6 @@ export default {
   updateLockName,
   updateLockSymbol,
   setBaseTokenURI,
+  addLockManager,
+  renounceLockManager,
 }