+++ packages/unlock-js/src/web3Service.ts
@@ -111,26 +111,55 @@ export default class Web3Service extends UnlockService {
     }
 
     const provider = this.providerForNetwork(network)
+
+    const version = await this.lockContractAbiVersion(
+      address,
+      this.providerForNetwork(network)
+    )
+
+    const lock = await version.getLock.bind(this)(
+      address,
+      this.providerForNetwork(network)
+    )
+    // Add the lock address
+    lock.address = address
+
+    lock.unlockContractAddress = ethers.utils.getAddress(
+      lock.unlockContractAddress
+    )
+
+    const previousDeployAddresses = (networkConfig.previousDeploys || []).map(
+      (d) => ethers.utils.getAddress(d.unlockAddress)
+    )
+    const isPreviousUnlockContract = previousDeployAddresses.includes(
+      lock.unlockContractAddress
+    )
+
+    const isUnlockContract =
+      ethers.utils.getAddress(networkConfig.unlockAddress) ===
+      lock.unlockContractAddress
+
+    // Check that the Unlock address matches one of the configured ones
+    if (!isUnlockContract && !isPreviousUnlockContract) {
+      throw new Error(
+        'This contract is not deployed from Unlock factory contract.'
+      )
+    }
+
+    // Check that the Unlock contract has indeed deployed this lock
     const unlockContract = await this.getUnlockContract(
-      networkConfig.unlockAddress,
+      lock.unlockContractAddress,
       provider
     )
 
     const response = await unlockContract.locks(address)
 
     if (!response.deployed) {
-      console.warn('Lock is not deployed from unlock factory contract.')
+      throw new Error(
+        'This contract is not deployed from Unlock factory contract.'
+      )
     }
 
-    const version = await this.lockContractAbiVersion(
-      address,
-      this.providerForNetwork(network)
-    )
-    const lock = await version.getLock.bind(this)(
-      address,
-      this.providerForNetwork(network)
-    )
-    lock.address = address
     return lock
   }